pyproject.toml
[tool.poetry]
name = "benchmark"
version = "0.1.0"
description = ""
authors = ["Washington Luiz Meireles de Lima <luiz.meireles@hotmail.co.uk>"]

[tool.poetry.dependencies]
python = "^3.10"
nats-py = "^2.1.4"
PyYAML = "^6.0"
black = "^22.6.0"
numpy = "^1.23.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
